requirements.txt
python-socketio
python-socketio[async]
aiohttp
requests
websocket-client